Conflicts:
	KittinCellProjectV0/src/Mocks/PartsRobotAgents/MockPartsRobot.java

+++ KittinCellProjectV0/src/Agents/PartsRobotAgent/NestAgent.java
@@ -4,39 +4,41 @@ import  Agent.*;
 import Agents.VisionAgent.VisionAgent;
 import server.Server;
 import data.*;
-import data.Part.PartType;
+
+
 import Interface.PartsRobotAgent.*;
+import Interface.VisionAgent.Vision;
 
 public class NestAgent extends Agent implements Nest{
 
 	//Data
-	Part.PartType parttype;
-	Part[] nestslots = new Part[9];
-	PartsRobotAgent partsrobot;
+	public PartInfo partinfo;
+	public Part[] nestslots = new Part[8];
+	PartsRobot partsrobot;
 	Lane lane;
-	VisionAgent camera;
+	Vision camera;
 	String name;
 	Server server;
 		
 	public int index;
 
-	private enum LaneStatus {hasPart, gettingParts, noAction}
-	private enum PartsRobotStatus {wantsParts, waitingForParts, noAction,readyforpart}
-	private enum NestStatus {badParts, noParts, needCheck, noAction}
-	private enum AnimationStatus {purging, needPurge, needSettle, noAction}
+	public enum LaneStatus {hasPart, gettingParts, noAction}
+	public enum PartsRobotStatus {wantsParts, waitingForParts, noAction,readyforpart}
+	public enum NestStatus {badParts, noParts, needCheck, noAction}
+	public enum AnimationStatus {purging, needPurge, needSettle, noAction}
 	
-	LaneStatus lanestate = LaneStatus.noAction;
-	PartsRobotStatus partsrobotstate= PartsRobotStatus.noAction;
-	NestStatus neststate = NestStatus.noParts;
-	AnimationStatus animationstate = AnimationStatus.noAction;
+	public LaneStatus lanestate = LaneStatus.noAction;
+	public PartsRobotStatus partsrobotstate= PartsRobotStatus.noAction;
+	public NestStatus neststate = NestStatus.noParts;
+	public AnimationStatus animationstate = AnimationStatus.noAction;
 	
-	public NestAgent(Lane lane, VisionAgent camera, int index)
+	public NestAgent(Lane lane, Vision camera, int index)
 	{
 		this.lane = lane;
 		this.camera = camera;
 		this.index = index;
 		name = "NestAgent " + index;
-		for(int i = 0; i<9; i++){
+		for(int i = 0; i<8; i++){
 			nestslots[i]=null;
 		}
 	}
@@ -48,12 +50,12 @@ public class NestAgent extends Agent implements Nest{
 		this.index = index;
 		this.server = server;
 		name = "NestAgent " + index;
-		for(int i = 0; i<9; i++){
+		for(int i = 0; i<8; i++){
 			nestslots[i]=null;
 		}
 	}	
 
-	public void setPartsRobotAgent(PartsRobotAgent robot){
+	public void setPartsRobotAgent(PartsRobot robot){
 		this.partsrobot = robot;
 	}
 	//Messages
@@ -64,7 +66,7 @@ public class NestAgent extends Agent implements Nest{
 	}
 	public void msgHereIsPart(Part p)
 	{
-			for(int i = 0; i < 9; i++)
+			for(int i = 0; i < 8; i++)
 			{
 				if(nestslots[i]== null)
 				{
@@ -74,6 +76,7 @@ public class NestAgent extends Agent implements Nest{
 					break;
 				}
 			}
+
 			stateChanged();
 	}
 
@@ -89,17 +92,17 @@ public class NestAgent extends Agent implements Nest{
 		stateChanged();
 	}
 
-	public void msgNeedThisPart(Part.PartType type)
+	public void msgNeedThisPart(PartInfo p)
 	{
-			for(int i = 0; i<9; i++)
+			for(int i = 0; i<8; i++)
 			{
 				if(nestslots[i] != null){
-					if(nestslots[i].type != type)
+					if(!nestslots[i].info.equals(partinfo))
 						animationstate = AnimationStatus.needPurge;
 				}
 			}
 					
-			parttype = type;
+			partinfo = p;
 			partsrobotstate = PartsRobotStatus.wantsParts;
 			
 			stateChanged();
@@ -120,6 +123,8 @@ public class NestAgent extends Agent implements Nest{
 	
 	public boolean pickAndExecuteAnAction()
 	{
+		
+		
 		if(partsrobotstate == PartsRobotStatus.readyforpart)
 		{
 			givePart();
@@ -130,7 +135,7 @@ public class NestAgent extends Agent implements Nest{
 			settleNest();
 			return true;
 		}
-		if(neststate == NestStatus.badParts || animationstate == AnimationStatus.needPurge)
+		if((neststate == NestStatus.badParts || animationstate == AnimationStatus.needPurge)&&animationstate != AnimationStatus.purging)
 		{
 			purgeNest();
 			return true;
@@ -143,7 +148,7 @@ public class NestAgent extends Agent implements Nest{
 		
 		if(lanestate == LaneStatus.hasPart)
 		{
-			for(int i = 0; i<9; i++)
+			for(int i = 0; i<8; i++)
 			{
 				if(nestslots[i] == null)
 				{
@@ -165,6 +170,7 @@ public class NestAgent extends Agent implements Nest{
 			settleNest();
 			return true;
 		}
+		print("current" + lanestate);
 		return false;
 	}
 
@@ -176,7 +182,7 @@ public class NestAgent extends Agent implements Nest{
 		//server.DoPurge(); // Run the purge animation
 		
 		animationstate = AnimationStatus.purging;
-		for(int i = 0; i<9; i++)
+		for(int i = 0; i<8; i++)
 		{
 			nestslots[i] = null;
 		}
@@ -184,6 +190,7 @@ public class NestAgent extends Agent implements Nest{
 
 	private void acceptPart()
 	{
+		
 		print("Ready to take the part");
 		lane.msgReadyForPart();
 		lanestate = LaneStatus.noAction;
@@ -200,15 +207,18 @@ public class NestAgent extends Agent implements Nest{
 	{
 		//NestSettlingAnimation
 		//gui.DoSettleNest(); // Settle Nest so that the parts behind in the nest move to the front
-		for(int i = 0; i<8; i++){
+		for(int i = 0; i<7; i++){
 			if(nestslots[i] == null && nestslots[i+1]!= null){
-				print("settling");
 				nestslots[i] = nestslots[i+1];
 				nestslots[i+1] = null;
 				if(i==0)
 					neststate = NestStatus.needCheck;
 			}
+			
 		}
+		//if(lanestate == LaneStatus.hasPart && nestslots[7]==null){
+		//	acceptPart();
+		//}
 	}
 	private void givePart()
 	{
@@ -216,14 +226,14 @@ public class NestAgent extends Agent implements Nest{
 		partsrobot.msgHereIsPart(nestslots[0]);
 		nestslots[0] = null;
 		partsrobotstate = PartsRobotStatus.waitingForParts;
-		//settleNest();
+		settleNest();
 	}
 		
 	private void askForParts()
 	{
 		print("Please give me a part");
 		if(lane!= null)
-		lane.msgNeedThisPart (parttype);
+		lane.msgNeedThisPart (partinfo);
 		partsrobotstate = PartsRobotStatus.waitingForParts;
 	}
 	
@@ -235,8 +245,8 @@ public class NestAgent extends Agent implements Nest{
 		return index;
 	}
 
-	public PartType getPartType() {
-		return parttype;
+	public PartInfo getPartInfo() {
+		return partinfo;
 	}
 	
 	public void setVisionAgent(VisionAgent camera){
@@ -245,6 +255,9 @@ public class NestAgent extends Agent implements Nest{
 	
 	public Integer getIndex() {
 		return index;
+	}
+	public void setLane(Lane newlane){
+		this.lane = newlane;
 	}
 	
 }
